longitude/latidtude as decimal numbers

+++ src/main/java/com/jinnova/smartpad/partner/IGPSInfo.java
@@ -1,12 +1,14 @@
 package com.jinnova.smartpad.partner;
 
+import java.math.BigDecimal;
+
 public interface IGPSInfo {
 	
-	float getLongitude();
+	BigDecimal getLongitude();
 	
-	void setLongitude(float f);
+	void setLongitude(BigDecimal f);
 	
-	float getLatitude();
+	BigDecimal getLatitude();
 	
-	void setLatitude(float f);
+	void setLatitude(BigDecimal f);
 }
